Reorganized files now that tests are passing
