+++ src/dom/patch.js
@@ -1,7 +1,7 @@
 import {setAttribute, removeAttribute} from './setAttribute'
-import {insertAtIndex} from './utils'
+import {insertAtIndex} from '../shared/utils'
 import createElement from './createElement'
-import {Actions, diffNode} from './diff'
+import {Actions, diffNode} from '../diff'
 
 /**
  * Modify a DOM element given an array of actions. A context can be set
