+++ src/shared/svg.js
